Got the results page to show dynamic content!!! V exciting!

+++ src/main/StudySpace.java
@@ -50,13 +50,13 @@ public class StudySpace {
 	public double getLongitude() {
 		return longitude;
 	}
-	public void setLongitude(float longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
 	public double getLatitude() {
 		return latitude;
 	}
-	public void setLatitude(float latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
 	public String getPhotoURL() {
